Unify VectorWriter with string writer, and move to separate file
