+++ src/common/vector/string_vector.cpp
@@ -5,11 +5,12 @@
 
 namespace duckdb {
 
-FlatVector::FlatStringWriter::FlatStringWriter(Vector &vector, idx_t count)
-    : vector(vector), data(GetDataMutable<string_t>(vector)), validity(Validity(vector)), count(count) {
+VectorWriter<string_t>::VectorWriter(Vector &vector, idx_t count)
+    : vector(vector), data(FlatVector::GetDataMutable<string_t>(vector)), validity(FlatVector::Validity(vector)),
+      count(count) {
 }
 
-void FlatVector::FlatStringWriter::InitializeHeap() {
+void VectorWriter<string_t>::InitializeHeap() {
 	heap = StringVector::GetStringHeap(vector);
 }
