+++ src/include/duckdb/common/vector/flat_vector.hpp
@@ -10,7 +10,6 @@
 
 #include "duckdb/common/types/vector.hpp"
 #include "duckdb/common/vector/constant_vector.hpp"
-#include "duckdb/common/types/string_heap.hpp"
 
 namespace duckdb {
 
@@ -62,6 +61,9 @@ protected:
 	idx_t capacity;
 	AllocatedData allocated_data;
 };
+
+template <class T>
+struct VectorWriter;
 
 struct FlatVector {
 	static void VerifyFlatVector(const Vector &vector) {
@@ -161,116 +163,9 @@ struct FlatVector {
 	}
 	DUCKDB_API static const SelectionVector *IncrementalSelectionVector();
 
-public:
 	template <class T>
-	struct FlatVectorWriter {
-		FlatVectorWriter(Vector &vector, idx_t count)
-		    : data(GetDataMutable<T>(vector)), validity(Validity(vector)), count(count) {
-		}
-
-		void SetInvalid(idx_t idx) {
-			D_ASSERT(idx < count);
-			validity.SetInvalid(idx);
-		}
-
-		T &operator[](idx_t idx) {
-			D_ASSERT(idx < count);
-			return data[idx];
-		}
-
-	private:
-		T *data;
-		ValidityMask &validity;
-		idx_t count;
-	};
-
-	struct FlatStringWriter;
-	struct StringElement {
-		StringElement(FlatStringWriter &writer, string_t *data, idx_t idx) : writer(writer), data(data), idx(idx) {
-		}
-
-		//! Constructs an empty string of a given length and returns it
-		//! Note: the empty string must be filled and .Finalize() must be called on it
-		inline string_t &EmptyString(idx_t length) {
-			if (length <= string_t::INLINE_LENGTH) {
-				data[idx] = string_t(UnsafeNumericCast<uint32_t>(length));
-			} else {
-				auto &heap = writer.GetHeap();
-				data[idx] = heap.CreateEmptyStringInHeap(length);
-			}
-			return data[idx];
-		}
-		inline string_t &operator=(string_t val) {
-			if (val.IsInlined()) {
-				data[idx] = val;
-			} else {
-				auto &heap = writer.GetHeap();
-				data[idx] = heap.AddBlobToHeap(val.GetData(), val.GetSize());
-			}
-			return data[idx];
-		}
-		inline void AssignWithoutCopying(string_t val) {
-			data[idx] = val;
-		}
-		inline char *GetDataWriteable() {
-			return data[idx].GetDataWriteable();
-		}
-		inline void Finalize() {
-			data[idx].Finalize();
-		}
-		inline string GetString() {
-			return data[idx].GetString();
-		}
-
-		operator string_t() const { // NOLINT: allow implicit conversion
-			return data[idx];
-		}
-
-	private:
-		FlatStringWriter &writer;
-		string_t *data;
-		idx_t idx;
-	};
-
-public:
-	struct FlatStringWriter {
-		FlatStringWriter(Vector &vector, idx_t count);
-
-		inline void SetInvalid(idx_t idx) {
-			D_ASSERT(idx < count);
-			validity.SetInvalid(idx);
-		}
-
-		inline StringElement operator[](idx_t idx) {
-			D_ASSERT(idx < count);
-			return StringElement(*this, data, idx);
-		}
-
-		inline StringHeap &GetHeap() {
-			if (!heap) {
-				InitializeHeap();
-			}
-			return *heap;
-		}
-
-	private:
-		void InitializeHeap();
-
-	private:
-		Vector &vector;
-		string_t *data;
-		ValidityMask &validity;
-		optional_ptr<StringHeap> heap;
-		idx_t count;
-	};
-
-	template <class T, typename std::enable_if<std::is_same<T, string_t>::value, int>::type = 0>
-	static FlatStringWriter Writer(Vector &vector, idx_t count) {
-		return FlatStringWriter(vector, count);
-	}
-	template <class T, typename std::enable_if<!std::is_same<T, string_t>::value, int>::type = 0>
-	static FlatVectorWriter<T> Writer(Vector &vector, idx_t count) {
-		return FlatVectorWriter<T>(vector, count);
+	static VectorWriter<T> Writer(Vector &vector, idx_t count) {
+		return VectorWriter<T>(vector, count);
 	}
 	template <class T>
 	static auto Writer(Vector &vector) -> decltype(Writer<T>(vector, NumericLimits<idx_t>::Maximum())) {
@@ -279,3 +174,5 @@ public:
 };
 
 } // namespace duckdb
+
+#include "duckdb/common/vector/vector_writer.hpp"
